requirements.txt
Jinja2==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.23
gunicorn==21.2.0
psycopg==3.1.18
waitress==3.0.0